Move main workflow to `core`, rm sql file
